py-openid-connect/requirements.txt
oauthlib
PyJWT
pyramid
pyramid_jinja2
pyramid_tm
requests[security]